+++ lib/shrine/plugins/direct_upload.rb
@@ -96,11 +96,11 @@ class Shrine
     #     plugin :direct_upload, presign_options: {acl: "public-read"}
     #
     #     plugin :direct_upload, presign_options: ->(request) do
-    #       filename = request.params["filename"].inspect
+    #       filename = request.params["filename"]
     #
     #       {
-    #         content_length_range: 0..(10*1024*1024),                 # limit filesize to 10MB
-    #         content_disposition: "attachment; filename=#{filename}", # download with original filename
+    #         content_length_range: 0..(10*1024*1024),                     # limit filesize to 10MB
+    #         content_disposition: "attachment; filename=\"#{filename}\"", # download with original filename
     #       }
     #     end
     #